Adding separate source and header for solver functions.
Removed test4 & test5 functions along with putting the solver functions in its own separate source and header files in order to clean up the code a bit.
